services/api/requirements.txt
# DESTINO: services/api/requirements.txt
# v2: aioredis→redis[asyncio] (aioredis 2.0.1 quebra no Python 3.11)

fastapi==0.110.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
h11==0.14.0
httptools==0.6.1

# Redis — substitui aioredis
redis[asyncio]>=4.6.0,<6.0.0

# Dados
polars==0.20.15
pyarrow==15.0.2

# Utilitários
python-decouple==3.8
structlog==24.1.0
